+++ airbyte-cdk/python/airbyte_cdk/sources/abstract_source.py
@@ -18,11 +18,11 @@ from airbyte_cdk.models import (
     SyncMode,
 )
 from airbyte_cdk.models import Type as MessageType
-from airbyte_cdk.sources.concurrent.full_refresh_stream_reader import FullRefreshStreamReader
-from airbyte_cdk.sources.concurrent.synchronous_full_refresh_reader import SyncrhonousFullRefreshReader
 from airbyte_cdk.sources.connector_state_manager import ConnectorStateManager
 from airbyte_cdk.sources.message import MessageRepository
 from airbyte_cdk.sources.source import Source
+from airbyte_cdk.sources.stream_reader.full_refresh_stream_reader import FullRefreshStreamReader
+from airbyte_cdk.sources.stream_reader.synchronous_full_refresh_reader import SyncrhonousFullRefreshReader
 from airbyte_cdk.sources.streams import Stream
 from airbyte_cdk.sources.streams.core import StreamData
 from airbyte_cdk.sources.streams.http.http import HttpStream
