+++ airbyte-cdk/python/airbyte_cdk/sources/stream_reader/__init__.py
@@ -1,0 +1,3 @@
+#
+# Copyright (c) 2023 Airbyte, Inc., all rights reserved.
+#
